moved app to its own seperate file
